feat: move formatAttribute function in AbstractCollectionRoute
